+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## Current (in progress)
 
-- Allow `Content-Type` and `Accept` in cors preflight headers
+- Allow more headers in cors preflight headers
   [#857](https://github.com/opendatateam/udata/pull/857)
+  [#860](https://github.com/opendatateam/udata/pull/860)
 
 ## 1.0.7 (2017-04-07)